task: Update daily functionality
