+++ utils/data/gamesData.js
@@ -1,0 +1,59 @@
+const gamesData = [
+    {
+        game: "AFK Journey",
+        server: "Global",
+        timezone: "Europe/Madrid",
+        dailyReset: "02:00",
+    },
+    {
+		game: "Arknights",
+		server: "EN",
+		timezone: "Etc/GMT+7",
+		dailyReset: "04:00",
+	},
+    {
+		game: "Azur Lane",
+		server: "EN",
+		timezone: "Etc/GMT+7",
+		dailyReset: "00:00",
+	},
+    {
+		game: "Blue Archive",
+		server: "Global",
+		timezone: "Etc/UTC",
+		dailyReset: "19:00",
+	},
+    {
+        game: "GODDESS OF VICTORY: NIKKE",
+        server: "Global",
+        timezone: "Etc/GMT-9",
+        dailyReset: "05:00",
+    },
+    {
+		game: "Honkai: Star Rail",
+		server: "America",
+		timezone: "Etc/GMT+5",
+		dailyReset: "04:00",
+	},
+    {
+		game: "Honkai Impact 3rd",
+		server: "Americas",
+		timezone: "Etc/GMT+5",
+		dailyReset: "04:00",
+	},
+    {
+		game: "Reverse: 1999",
+		server: "Global",
+		timezone: "Etc/GMT+5",
+		dailyReset: "05:00",
+	},
+    {
+        game: "Solo Leveling",
+        server: "Global",
+        timezone: "Europe/Madrid",
+        dailyReset: "02:00",
+    },
+    // TODO: Add more games as needed. REF: https://raw.githubusercontent.com/cicerakes/Game-Time-Master/master/game-data.js
+];
+
+module.exports = gamesData;
