+++ utils/index.js
@@ -2,6 +2,7 @@ const getFiles = require('./getFiles');
 const { getIsStreaming, setIsStreaming } = require('./getActivity');
 const getRapiMessages = require("./getRapiMessages");
 const { getBosses, getBossesLinks, getTribeTowerRotation, getBossFileName } = require("./getDailyInterceptionConfig");
+const gamesData = require('./data/gamesData');
 
 module.exports = {
 	getFiles,
@@ -12,4 +13,5 @@ module.exports = {
     getBossesLinks,
     getTribeTowerRotation,
     getBossFileName,
+    gamesData
 }
